INTENT_NETWORKING/requirements.txt
PyYAML==6.0
junos-eznc==2.6.7
ncclient==0.6.13
paramiko==3.1.0
#pip install -r requirements.txt